requirements.txt
openapi
termcolor
python-dotenv
